src/main/java/io/gdcc/spi/export/debug/DebugExporter.java
package io.gdcc.spi.export.debug;

import com.google.auto.service.AutoService;

import io.gdcc.spi.export.ExportDataProvider;
import io.gdcc.spi.export.ExportException;
import io.gdcc.spi.export.Exporter;

import jakarta.json.Json;
import jakarta.json.JsonArray;
import jakarta.json.JsonObject;
import jakarta.json.JsonObjectBuilder;
import jakarta.ws.rs.core.MediaType;

import java.io.OutputStream;
import java.util.Locale;

@AutoService(Exporter.class)
public class DebugExporter implements Exporter {

    /**
     * The name of the format it creates. If this format is already provided by a built-in exporter,
     * this Exporter will override the built-in one. (Note that exports are cached, so existing
     * metadata export files are not updated immediately.)
     */
    @Override
    public String getFormatName() {
        return "debug";
    }

    /**
     * The display name shown in the UI.
     *
     * @param locale Used to generate a translation.
     */
    @Override
    public String getDisplayName(Locale locale) {
        return "Debug";
    }

    /** Whether the exported format should be available as an option for Harvesting. */
    @Override
    public Boolean isHarvestable() {
        return false;
    }

    /** Whether the exported format should be available for download in the UI and API. */
    @Override
    public Boolean isAvailableToUsers() {
        return true;
    }

    /**
     * Defines the mime type of the exported format. Used when metadata is downloaded, i.e. to
     * trigger an appropriate viewer in the user's browser.
     */
    @Override
    public String getMediaType() {
        return MediaType.APPLICATION_JSON;
    }

    /**
     * This method is called by Dataverse when metadata for a given dataset in this format is
     * requested.
     */
    @Override
    public void exportDataset(ExportDataProvider dataProvider, OutputStream outputStream)
            throws ExportException {
        try {
            JsonObject datasetJson = dataProvider.getDatasetJson();
            JsonObject datasetORE = dataProvider.getDatasetORE();
            JsonObject datasetSchemaDotOrg = dataProvider.getDatasetSchemaDotOrg();
            JsonArray datasetFileDetails = dataProvider.getDatasetFileDetails();
            String dataCiteXml = dataProvider.getDataCiteXml();

            JsonObjectBuilder job = Json.createObjectBuilder();
            job.add("datasetJson", datasetJson);
            job.add("datasetORE", datasetORE);
            job.add("datasetSchemaDotOrg", datasetSchemaDotOrg);
            job.add("datasetFileDetails", datasetFileDetails);
            job.add("dataCiteXml", dataCiteXml);

            // Write the output format to the output stream.
            outputStream.write(job.build().toString().getBytes("UTF8"));
            // Flush the output stream - The output stream is automatically closed by
            // Dataverse and should not be closed in the Exporter.
            outputStream.flush();
        } catch (Exception ex) {
            // If anything goes wrong, an Exporter should throw an ExportException.
            throw new ExportException("Unknown exception caught during export: " + ex);
        }
    }
}
